Add wrap around for long text rows.
Added config to store a max_width data point. Buffer considers this and,
if inserting a char to a line that will overflow, makes a new line, then
returns that it did so to let the command know. Very imperfect. Two
flaws: 1. Deleting after wrapping around to a new line is bugging, can't
figure out why yet. 2. The max_width does not actually fill up the
terminal.